+++ apps/consent-ui/src/app/[lang]/register/RegistrationForm.tsx
@@ -23,7 +23,7 @@ import { SyntheticEvent, useState } from 'react';
 import axios from 'axios';
 import urlJoin from 'url-join';
 import Button from 'src/components/Button';
-import useRecaptcha from 'src/components/RecaptchaCheckbox/useRecaptcha';
+import useRecaptcha from 'src/hooks/useRecaptcha';
 import RecaptchaCheckbox from 'src/components/RecaptchaCheckbox';
 import { useAppConfigContext } from 'src/components/AppConfigContextProvider';
 import Notification from 'src/components/Notification';
