+++ apps/consent-ui/src/components/RecaptchaCheckbox/index.tsx
@@ -24,7 +24,7 @@ import ReCAPTCHA from 'react-google-recaptcha';
 
 import { useAppConfigContext } from '../AppConfigContextProvider';
 
-import { RecaptchaCheckboxRef } from './useRecaptcha';
+import { RecaptchaCheckboxRef } from '../../hooks/useRecaptcha';
 
 const RecaptchaCheckbox = ({
 	onChange,
